requirements.txt
django>=3.2.11,<3.3
django-cors-headers
django_crispy_forms
django-extensions
django-heroku
django-simple-history
django-localflavor
django-admin-list-filter-dropdown
django-smart-selects
djangorestframework
gunicorn
psycopg2
psycopg2-binary
whitenoise